pyproject.toml
[project]
name = "recurse-rlm"
version = "0.1.0"
requires-python = ">=3.11"
license = { text = "MIT" }
description = "RLM-powered unlimited context MCP server using Qwen 3.5 via Ollama"

dependencies = [
    "openai>=1.14.0",
    "mcp>=1.0.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
docker = ["docker>=7.0.0"]

[project.scripts]
recurse = "recurse.server:main"

[tool.hatch.build.targets.wheel]
packages = ["recurse"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
